Dataset: Renames normal Dataset to ReactionDataset and begins to subclass the dataset object
