+++ qcfractal/interface/collections/__init__.py
@@ -4,5 +4,6 @@ Python init for QCPortal collections
 
 from .collection_utils import collection_factory
 from .dataset import Dataset
+from .reaction_dataset import ReactionDataset
 from .generic import Generic
 from .openffworkflow import OpenFFWorkflow
